Add HBaseHealthDetector Application
