+++ eagle-dev/checkstyle.xml
@@ -33,7 +33,7 @@
 <module name="Checker">
     <property name="charset" value="UTF-8"/>
 
-    <property name="status" value="warning"/>
+    <property name="severity" value="warning"/>
 
     <property name="fileExtensions" value="java, properties, xml"/>
