+++ eagle-hadoop-health/src/main/java/org/apache/eagle/health/jobs/URLPingCheckJob.java
@@ -14,11 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.eagle.alert.engine.coordinator;
+package org.apache.eagle.health.jobs;
 
-/**
- * Severity Levels for alerts, events, logs, etc.
- */
-public enum AlertSeverity {
-    UNKNOWN, OK, WARNING, CRITICAL, FATAL
+public class URLPingCheckJob {
+
 }
