+++ eagle-hadoop-health/src/test/resources/quartz.properties
