--- eagle-hadoop-servicecheck/src/main/resources/META-INF/providers/org.apache.eagle.servicecheck.HadoopServiceCheckAppProvider.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-  ~ Licensed to the Apache Software Foundation (ASF) under one or more
-  ~ contributor license agreements.  See the NOTICE file distributed with
-  ~ this work for additional information regarding copyright ownership.
-  ~ The ASF licenses this file to You under the Apache License, Version 2.0
-  ~ (the "License"); you may not use this file except in compliance with
-  ~ the License.  You may obtain a copy of the License at
-  ~
-  ~    http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<application>
-    <type>HADOOP_HEALTHCHECK_APP</type>
-    <name>Hadoop Service HealthCheck</name>
-    <configuration>
-    </configuration>
-    <streams>
-    </streams>
-    <docs>
-        <install>
-            <b>How to Install</b>
-            TODO
-        </install>
-        <uninstall>
-            <b>How to Uninstall</b>
-            TODO
-        </uninstall>
-    </docs>
-</application>
